chore: add utility functions for props handling
