+++ www/components/mdx/props-table.tsx
@@ -2,7 +2,7 @@ import type { Props } from "@/data"
 import { Box, Code, Grid, NativeTable, Text } from "@yamada-ui/react"
 import { useTranslations } from "next-intl"
 import { Fragment } from "react"
-import { getProps, isNamespace, isSingle } from "@/data"
+import { getProps, isNamespace, isSingle } from "@/utils/props"
 import { mdToHtml } from "@/utils/string"
 import { H3 } from "./heading"
 import { Link } from "./link"
